Update scraper for current web page
Update python script to find the correct class in the html of the UCI
website. Im guessing the class name changed since the last update of
this scraper.

Current version of the scraper successfully finds the web page and
pulls the html with beautifulsoup, but the find statement returns empty
because the class it looks for does not exist.

I didnt find any changes in the actual requirements, but formatting
changes (newline locations) in the files written by the scraper
necessitate rewriting of the txt files.