Update scraper and current scrape (#14)
* Add switch for exclusion of CV section

The CV section is not needed for a master's thesis. This adds a switch to the preliminarypages function which excludes the CV section if it is commented out in the preliminaries file.

* Add comment to CV section

* Add field to change the document title

The title of the document should be dissertation for PhD, but thesis for master's. This adds an input for the document title in preliminaries.tex and inserts that title into the correct places in the document.

* Update scraper and current scrape. No significant changes to manual.

Update the URLs in the scraper script to work with the current version
of the manual website. Also update the html class BeautifulSoup looks
for.

Only cosmetic changes were found to the manual, so no changes to the
format are necessary.